app/client.py
from __future__ import print_function
import grpc
import api.gen.helloworld_pb2 as helloworld_pb2
import api.gen.helloworld_pb2_grpc as helloworld_pb2_grpc


def run():
    channel = grpc.insecure_channel('localhost:50051')
    stub = helloworld_pb2_grpc.GreeterStub(channel)
    response = stub.SayHello(helloworld_pb2.HelloRequest(name='Hello World！ This is message from client!'))
    print("Greeter client received:" + response.message)


if __name__ == '__main__':
    run()
